PR fixes

+++ src/main/java/com/arpnetworking/kairosdb/aggregators/MovingWindowAggregator.java
@@ -406,7 +406,7 @@ public class MovingWindowAggregator extends RangeAggregator {
         }
 
         @Override
-        public int getSampleCount() {
+        public long getSampleCount() {
             return _wrappedDataPoint.getSampleCount();
         }
